backend-admin/pyproject.toml
[project]
name = "lativ"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.2",
    "boto3==1.40.17",
    "fastapi[standard]>=0.116.1",
    "mypy-boto3-builder>=8.11.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "boto3-stubs[s3]>=1.40.18",
    "jinja2>=3.1.6",
    "ruff>=0.12.10",
]
